// apps/storyblok-website-intl/src/lib/storyblok.types.ts
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
import type { ISbStoryData } from "storyblok";
/* eslint-disable -- This file was automatically generated by json-schema-to-typescript. -- DO NOT MODIFY IT BY HAND. Instead, modify the source JSON Schema file, -- and run json-schema-to-typescript to regenerate this file. */

export interface HeroSectionStoryblok {
  title?: string;
  description?: string;
  component: "HeroSection";
  _uid: string;
}

/* eslint-disable -- This file was automatically generated by json-schema-to-typescript. -- DO NOT MODIFY IT BY HAND. Instead, modify the source JSON Schema file, -- and run json-schema-to-typescript to regenerate this file. */

export interface PageStoryblok {
  body?: (HeroSectionStoryblok | PageStoryblok)[];
  title?: string;
  description?: string;
  isNoIndex?: boolean;
  ogTitle?: string;
  ogDescription?: string;
  ogType?: string;
  ogUrl?: string;
  component: "page";
  _uid: string;
}